refactor: extract input settings helpers
